Even more reorganization of the engine config UI code.
