+++ Source/Engines/ModuleEngineConfigs.cs
@@ -159,7 +159,7 @@ namespace RealFuels
             return node.GetValue(PatchNameKey); // Just show subconfig name without parent prefix
         }
 
-        protected override IEnumerable<ConfigRowDefinition> BuildConfigRows()
+        public override IEnumerable<ConfigRowDefinition> BuildConfigRows()
         {
             foreach (var node in FilteredDisplayConfigs(false))
             {
@@ -197,11 +197,6 @@ namespace RealFuels
                 }
             }
         }
-
-        protected override void DrawConfigSelectors(IEnumerable<ConfigNode> availableConfigNodes)
-        {
-            DrawConfigTable(BuildConfigRows());
-        }
     }
 
     public class ModuleEngineConfigsBase : PartModule, IPartCostModifier, IPartMassModifier
@@ -210,7 +205,7 @@ namespace RealFuels
         public const string groupName = ModuleEnginesRF.groupName;
         public const string groupDisplayName = "#RF_Engine_EngineConfigs"; // "Engine Configs"
         #region Fields
-        protected bool compatible = true;
+        internal bool compatible = true;
 
         [KSPField(isPersistant = true)]
         public string configuration = string.Empty;
@@ -361,7 +356,7 @@ namespace RealFuels
             field.group = new BasePAWGroup(groupName, groupDisplayName, false);
         }
 
-        protected List<ConfigNode> FilteredDisplayConfigs(bool update)
+        internal List<ConfigNode> FilteredDisplayConfigs(bool update)
         {
             if (update || filteredDisplayConfigs == null)
             {
@@ -821,7 +816,7 @@ namespace RealFuels
         }
 
         #region SetConfiguration Tools
-        private Dictionary<string, Gimbal> ExtractGimbals(ConfigNode cfg)
+        internal Dictionary<string, Gimbal> ExtractGimbals(ConfigNode cfg)
         {
             Gimbal ExtractGimbalKeys(ConfigNode c)
             {
@@ -1138,44 +1133,6 @@ namespace RealFuels
             // are shared across all instances. They'll be cleaned up when Unity unloads the scene.
         }
 
-        private static Vector3 mousePos = Vector3.zero;
-        private static Rect guiWindowRect = new Rect(0, 0, 0, 0);
-        private static uint lastPartId = 0;
-        private static int lastConfigCount = 0;
-        private static bool lastCompactView = false;
-        private static bool lastHasChart = false;
-        private string myToolTip = string.Empty;
-        private int counterTT;
-        private bool editorLocked = false;
-
-        private Vector2 configScrollPos = Vector2.zero;
-        private GUIContent configGuiContent;
-        private bool compactView = false;
-        private bool useLogScaleX = false; // Toggle for logarithmic x-axis on failure chart
-        private bool useLogScaleY = false; // Toggle for logarithmic y-axis on failure chart
-
-        // Column visibility customization
-        private bool showColumnMenu = false;
-        private static Rect columnMenuRect = new Rect(100, 100, 280, 650); // Separate window rect - tall enough for all columns
-        private static bool[] columnsVisibleFull = new bool[18];
-        private static bool[] columnsVisibleCompact = new bool[18];
-        private static bool columnVisibilityInitialized = false;
-
-        // Simulation controls for data percentage and cluster size
-        private bool useSimulatedData = false; // Whether to override real TestFlight data
-        private float simulatedDataValue = 0f; // Simulated data value in du (data units)
-        private int clusterSize = 1; // Number of engines in cluster (default 1)
-        private string clusterSizeInput = "1"; // Text input for cluster size
-        private string dataValueInput = "0"; // Text input for data value in du
-
-        private const int ConfigRowHeight = 22;
-        private const int ConfigMaxVisibleRows = 16; // Max rows before scrolling (60% taller)
-        // Dynamic column widths - calculated based on content
-        private float[] ConfigColumnWidths = new float[18];
-
-        // Texture and style management - using singletons to prevent memory leaks
-        private EngineConfigTextures Textures => EngineConfigTextures.Instance;
-
         // Tech level management - lazy initialization
         private EngineConfigTechLevels _techLevels;
         protected EngineConfigTechLevels TechLevels
@@ -1200,244 +1157,22 @@ namespace RealFuels
             }
         }
 
-        // Chart rendering - lazy initialization
-        private EngineConfigChart _chart;
-        private EngineConfigChart Chart
+        // GUI rendering - lazy initialization
+        private EngineConfigGUI _gui;
+        private EngineConfigGUI GUI
         {
             get
             {
-                if (_chart == null)
-                {
-                    _chart = new EngineConfigChart(this);
-                    _chart.UseLogScaleX = useLogScaleX;
-                    _chart.UseLogScaleY = useLogScaleY;
-                    _chart.UseSimulatedData = useSimulatedData;
-                    _chart.SimulatedDataValue = simulatedDataValue;
-                    _chart.ClusterSize = clusterSize;
-                }
-                return _chart;
+                if (_gui == null)
+                    _gui = new EngineConfigGUI(this);
+                return _gui;
             }
         }
 
-        private int toolTipWidth => EditorLogic.fetch.editorScreen == EditorScreen.Parts ? 220 : 300;
-        private int toolTipHeight => (int)Styles.styleEditorTooltip.CalcHeight(new GUIContent(myToolTip), toolTipWidth);
-
-        public void OnGUI()
-        {
-            if (!compatible || !isMaster || !HighLogic.LoadedSceneIsEditor || EditorLogic.fetch == null)
-                return;
-
-            bool inPartsEditor = EditorLogic.fetch.editorScreen == EditorScreen.Parts;
-            if (!(showRFGUI && inPartsEditor) && !(EditorLogic.fetch.editorScreen == EditorScreen.Actions && EditorActionGroups.Instance.GetSelectedParts().Contains(part)))
-            {
-                EditorUnlock();
-                return;
-            }
-
-            if (inPartsEditor && part.symmetryCounterparts.FirstOrDefault(p => p.persistentId < part.persistentId) is Part)
-                return;
-
-            if (guiWindowRect.width == 0)
-            {
-                int posAdd = inPartsEditor ? 256 : 0;
-                int posMult = (offsetGUIPos == -1) ? (part.Modules.Contains("ModuleFuelTanks") ? 1 : 0) : offsetGUIPos;
-                // Set position with minimal initial size - both width and height will auto-size tightly
-                guiWindowRect = new Rect(posAdd + 430 * posMult, 365, 100, 100);
-            }
-
-            // Only reset height when switching parts or when content changes (compact view, config count, chart visibility)
-            // This prevents flickering during dragging and slider interaction
-            uint currentPartId = part.persistentId;
-            int currentConfigCount = FilteredDisplayConfigs(false).Count;
-            bool currentHasChart = config != null && config.HasValue("cycleReliabilityStart");
-            bool contentChanged = currentPartId != lastPartId
-                               || currentConfigCount != lastConfigCount
-                               || compactView != lastCompactView
-                               || currentHasChart != lastHasChart;
-
-            if (contentChanged)
-            {
-                float savedX = guiWindowRect.x;
-                float savedY = guiWindowRect.y;
-                float savedWidth = guiWindowRect.width;
-                guiWindowRect = new Rect(savedX, savedY, savedWidth, 100);
-
-                lastPartId = currentPartId;
-                lastConfigCount = currentConfigCount;
-                lastCompactView = compactView;
-                lastHasChart = currentHasChart;
-            }
-
-            mousePos = Input.mousePosition; //Mouse location; based on Kerbal Engineer Redux code
-            mousePos.y = Screen.height - mousePos.y;
-            if (guiWindowRect.Contains(mousePos))
-                EditorLock();
-            else
-                EditorUnlock();
-
-            myToolTip = myToolTip.Trim();
-            if (!string.IsNullOrEmpty(myToolTip))
-            {
-                int offset = inPartsEditor ? -222 : 440;
-                GUI.Label(new Rect(guiWindowRect.xMin + offset, mousePos.y - 5, toolTipWidth, toolTipHeight), myToolTip, Styles.styleEditorTooltip);
-            }
-
-            guiWindowRect = GUILayout.Window(unchecked((int)part.persistentId), guiWindowRect, EngineManagerGUI, Localizer.Format("#RF_Engine_WindowTitle", part.partInfo.title), Styles.styleEditorPanel); // "Configure " + part.partInfo.title
-
-            // Draw column menu as separate window if open
-            if (showColumnMenu)
-            {
-                columnMenuRect = GUI.Window(unchecked((int)part.persistentId) + 1, columnMenuRect, DrawColumnMenuWindow, "Settings", HighLogic.Skin.window);
-            }
-        }
-
-        private void DrawColumnMenuWindow(int windowID)
-        {
-            DrawColumnMenu(new Rect(0, 20, columnMenuRect.width, columnMenuRect.height - 20));
-            GUI.DragWindow(new Rect(0, 0, columnMenuRect.width, 20));
-        }
-
-        private void EditorLock()
-        {
-            if (!editorLocked)
-            {
-                EditorLogic.fetch.Lock(false, false, false, "RFGUILock");
-                editorLocked = true;
-                KSP.UI.Screens.Editor.PartListTooltipMasterController.Instance?.HideTooltip();
-            }
-        }
-
-        private void EditorUnlock()
-        {
-            if (editorLocked)
-            {
-                EditorLogic.fetch.Unlock("RFGUILock");
-                editorLocked = false;
-            }
-        }
-
-        protected string GetCostString(ConfigNode node)
-        {
-            string costString = string.Empty;
-            if (node.HasValue("cost"))
-            {
-                float curCost = scale * float.Parse(node.GetValue("cost"), CultureInfo.InvariantCulture);
-
-                if (techLevel != -1)
-                {
-                    curCost = TechLevels.CostTL(curCost, node) - TechLevels.CostTL(0f, node); // get purely the config cost difference
-                }
-                costString = $" ({((curCost < 0) ? string.Empty : "+")}{curCost:N0}√)";
-            }
-            return costString;
-        }
-
-        /// Normal apply action for the 'select <config>' button.
-        protected void GUIApplyConfig(string configName)
-        {
-            SetConfiguration(configName, true);
-            UpdateSymmetryCounterparts();
-            GameEvents.onEditorShipModified.Fire(EditorLogic.fetch.ship);
-            MarkWindowDirty();
-        }
-
-        protected void DrawSelectButton(ConfigNode node, bool isSelected, Action<string> apply)
-        {
-            var nName = node.GetValue("name");
-            var dispName = GetConfigDisplayName(node);
-            var costString = GetCostString(node);
-            var configInfo = GetConfigInfo(node);
-
-            using (new GUILayout.HorizontalScope())
-            {
-                // For simulations, RP-1 will allow selecting all configs despite tech status or whether the entry cost has been paid.
-                // The KCT that comes with RP-1 will call the Validate() method when the player tries to add a vessel to the build queue.
-                if (Utilities.RP1Found)
-                {
-                    // Currently selected.
-                    if (isSelected)
-                    {
-                        GUILayout.Label(new GUIContent($"{Localizer.GetStringByTag("#RF_Engine_CurrentConfig")}: {dispName}{costString}", configInfo)); // Current config
-                    }
-                    else if (GUILayout.Button(new GUIContent($"{Localizer.GetStringByTag("#RF_Engine_ConfigSwitch")} {dispName}{costString}", configInfo))) // Switch to
-                        apply(nName);
-
-                    if (!EngineConfigTechLevels.UnlockedConfig(node, part))
-                    {
-                        double upgradeCost = EntryCostManager.Instance.ConfigEntryCost(nName);
-                        string techRequired = node.GetValue("techRequired");
-                        if (upgradeCost <= 0)
-                        {
-                            // Auto-buy.
-                            EntryCostManager.Instance.PurchaseConfig(nName, techRequired);
-                        }
-
-                        bool isConfigAvailable = EngineConfigTechLevels.CanConfig(node);
-                        string tooltip = string.Empty;
-                        if (!isConfigAvailable && techNameToTitle.TryGetValue(techRequired, out string techStr))
-                        {
-                            tooltip = Localizer.Format("#RF_Engine_LacksTech", techStr); // $"Lacks tech for {techStr}"
-                        }
-
-                        GUI.enabled = isConfigAvailable;
-                        if (GUILayout.Button(new GUIContent($"{Localizer.GetStringByTag("#RF_Engine_Purchase")} ({upgradeCost:N0}√)", tooltip), GUILayout.Width(145))) // Purchase
-                        {
-                            if (EntryCostManager.Instance.PurchaseConfig(nName, node.GetValue("techRequired")))
-                                apply(nName);
-                        }
-                        GUI.enabled = true;
-                    }
-                }
-                else
-                {
-                    // Currently selected.
-                    if (isSelected)
-                    {
-                        GUILayout.Label(new GUIContent($"{Localizer.GetStringByTag("#RF_Engine_CurrentConfig")}: {dispName}{costString}", configInfo)); // Current config
-                        return;
-                    }
-
-                    // Locked.
-                    if (!EngineConfigTechLevels.CanConfig(node))
-                    {
-                        if (techNameToTitle.TryGetValue(node.GetValue("techRequired"), out string techStr))
-                            techStr = $"\n{Localizer.GetStringByTag("#RF_Engine_Requires")}: " + techStr; // Requires
-                        GUILayout.Label(new GUIContent(Localizer.Format("#RF_Engine_LacksTech", dispName), configInfo + techStr)); // $"Lacks tech for {dispName}"
-                        return;
-                    }
-
-                    // Available.
-                    if (EngineConfigTechLevels.UnlockedConfig(node, part))
-                    {
-                        if (GUILayout.Button(new GUIContent($"{Localizer.GetStringByTag("#RF_Engine_ConfigSwitch")} {dispName}{costString}", configInfo))) // Switch to
-                            apply(nName);
-                        return;
-                    }
-
-                    // Purchase.
-                    double upgradeCost = EntryCostManager.Instance.ConfigEntryCost(nName);
-                    string techRequired = node.GetValue("techRequired");
-                    if (upgradeCost > 0d)
-                    {
-                        costString = $" ({upgradeCost:N0}√)";
-                        if (GUILayout.Button(new GUIContent($"{Localizer.GetStringByTag("#RF_Engine_Purchase")}  {dispName}{costString}", configInfo))) // Purchase
-                        {
-                            if (EntryCostManager.Instance.PurchaseConfig(nName, techRequired))
-                                apply(nName);
-                        }
-                    }
-                    else
-                    {
-                        // Auto-buy.
-                        EntryCostManager.Instance.PurchaseConfig(nName, techRequired);
-                        if (GUILayout.Button(new GUIContent($"{Localizer.GetStringByTag("#RF_Engine_ConfigSwitch")}  {dispName}{costString}", configInfo))) // Switch to
-                            apply(nName);
-                    }
-                }
-            }
-        }
-
-        protected struct ConfigRowDefinition
+        /// <summary>
+        /// Struct for passing configuration row data to GUI
+        /// </summary>
+        public struct ConfigRowDefinition
         {
             public ConfigNode Node;
             public string DisplayName;
@@ -1446,9 +1181,13 @@ namespace RealFuels
             public Action Apply;
         }
 
-        protected virtual IEnumerable<ConfigRowDefinition> BuildConfigRows()
+        /// <summary>
+        /// Builds the list of configuration rows to display in the GUI.
+        /// Virtual so derived classes can customize the row structure.
+        /// </summary>
+        public virtual IEnumerable<ConfigRowDefinition> BuildConfigRows()
         {
-            foreach (ConfigNode node in FilteredDisplayConfigs(false))
+            foreach (var node in FilteredDisplayConfigs(false))
             {
                 string configName = node.GetValue("name");
                 yield return new ConfigRowDefinition
@@ -1462,1009 +1201,43 @@ namespace RealFuels
             }
         }
 
-        protected void CalculateColumnWidths(List<ConfigRowDefinition> rows)
+        /// <summary>
+        /// Draws the configuration selector UI elements.
+        /// Virtual so derived classes can add custom UI before the config table.
+        /// </summary>
+        protected internal virtual void DrawConfigSelectors(IEnumerable<ConfigNode> availableConfigNodes)
         {
-            // Create style for measuring cell content
-            var cellStyle = new GUIStyle(GUI.skin.label)
-            {
-                fontSize = 14,
-                fontStyle = FontStyle.Bold,
-                padding = new RectOffset(5, 0, 0, 0)
-            };
-
-            // Initialize with minimum widths
-            for (int i = 0; i < ConfigColumnWidths.Length; i++)
-            {
-                ConfigColumnWidths[i] = 30f; // Start with minimum
-            }
-
-            // Measure all row content (ignore headers - they're rotated and centered)
-            foreach (var row in rows)
-            {
-                string nameText = row.DisplayName;
-                if (row.Indent) nameText = "    ↳ " + nameText;
-
-                string[] cellValues = new string[]
-                {
-                    nameText,
-                    GetThrustString(row.Node),
-                    GetMinThrottleString(row.Node),
-                    GetIspString(row.Node),
-                    GetMassString(row.Node),
-                    GetGimbalString(row.Node),
-                    GetIgnitionsString(row.Node),
-                    GetBoolSymbol(row.Node, "ullage"),
-                    GetBoolSymbol(row.Node, "pressureFed"),
-                    GetRatedBurnTimeString(row.Node),
-                    GetTestedBurnTimeString(row.Node), // NEW: Tested burn time column
-                    GetIgnitionReliabilityStartString(row.Node),
-                    GetIgnitionReliabilityEndString(row.Node),
-                    GetCycleReliabilityStartString(row.Node),
-                    GetCycleReliabilityEndString(row.Node),
-                    GetTechString(row.Node),
-                    GetCostDeltaString(row.Node),
-                    "" // Action column - buttons
-                };
-
-                for (int i = 0; i < cellValues.Length; i++)
-                {
-                    if (!string.IsNullOrEmpty(cellValues[i]))
-                    {
-                        float width = cellStyle.CalcSize(new GUIContent(cellValues[i])).x + 10f; // Add padding
-                        if (width > ConfigColumnWidths[i])
-                            ConfigColumnWidths[i] = width;
-                    }
-                }
-            }
-
-            // Action column needs fixed width for two buttons
-            ConfigColumnWidths[17] = 160f;
-
-            // Set minimum widths for specific columns
-            ConfigColumnWidths[7] = Mathf.Max(ConfigColumnWidths[7], 30f); // Ull
-            ConfigColumnWidths[8] = Mathf.Max(ConfigColumnWidths[8], 30f); // PFed
-            ConfigColumnWidths[9] = Mathf.Max(ConfigColumnWidths[9], 50f); // Rated burn
-            ConfigColumnWidths[10] = Mathf.Max(ConfigColumnWidths[10], 50f); // Tested burn
-        }
-
-        protected void DrawConfigTable(IEnumerable<ConfigRowDefinition> rows)
-        {
-            EnsureTexturesAndStyles();
-
-            var rowList = rows.ToList();
-
-            // Calculate dynamic column widths
-            CalculateColumnWidths(rowList);
-
-            // Sum only visible column widths
-            float totalWidth = 0f;
-            for (int i = 0; i < ConfigColumnWidths.Length; i++)
-            {
-                if (IsColumnVisible(i))
-                    totalWidth += ConfigColumnWidths[i];
-            }
-
-            // Update window width to fit table exactly (accounting for window padding: 5px left + 5px right = 10px)
-            float requiredWindowWidth = totalWidth + 10f; // Table width + padding
-            const float minWindowWidth = 900f; // Minimum width to prevent squishing
-            guiWindowRect.width = Mathf.Max(requiredWindowWidth, minWindowWidth);
-
-            Rect headerRowRect = GUILayoutUtility.GetRect(GUIContent.none, GUI.skin.label, GUILayout.Height(45));
-            float headerStartX = headerRowRect.x; // No left margin
-            DrawHeaderRow(new Rect(headerStartX, headerRowRect.y, totalWidth, headerRowRect.height));
-
-            // Dynamic height: grow up to max, then scroll
-            int actualRows = rowList.Count;
-            int visibleRows = Mathf.Min(actualRows, ConfigMaxVisibleRows);
-            int scrollViewHeight = visibleRows * ConfigRowHeight;
-
-            // No spacing in scroll view
-            var scrollStyle = new GUIStyle(GUI.skin.scrollView) { padding = new RectOffset(0, 0, 0, 0) };
-            configScrollPos = GUILayout.BeginScrollView(configScrollPos, false, false, GUIStyle.none, GUI.skin.verticalScrollbar, scrollStyle, GUILayout.Height(scrollViewHeight));
-
-            // Use a style with no margin/padding for tight row spacing
-            var noSpaceStyle = new GUIStyle { margin = new RectOffset(0, 0, 0, 0), padding = new RectOffset(0, 0, 0, 0) };
-
-            int rowIndex = 0;
-            foreach (var row in rowList)
-            {
-                Rect rowRect = GUILayoutUtility.GetRect(GUIContent.none, noSpaceStyle, GUILayout.Height(ConfigRowHeight));
-                float rowStartX = rowRect.x; // No left margin
-                Rect tableRowRect = new Rect(rowStartX, rowRect.y, totalWidth, rowRect.height);
-                bool isHovered = tableRowRect.Contains(Event.current.mousePosition);
-
-                bool isLocked = !EngineConfigTechLevels.CanConfig(row.Node);
-                if (Event.current.type == EventType.Repaint)
-                {
-                    // Draw alternating row background first
-                    if (!row.IsSelected && !isLocked && !isHovered && rowIndex % 2 == 1)
-                    {
-                        GUI.DrawTexture(tableRowRect, Textures.ZebraStripe);
-                    }
-
-                    if (row.IsSelected)
-                        GUI.DrawTexture(tableRowRect, Textures.RowCurrent);
-                    else if (isLocked)
-                        GUI.DrawTexture(tableRowRect, Textures.RowLocked);
-                    else if (isHovered)
-                        GUI.DrawTexture(tableRowRect, Textures.RowHover);
-                }
-
-                string tooltip = GetRowTooltip(row.Node);
-                if (configGuiContent == null)
-                    configGuiContent = new GUIContent();
-                configGuiContent.text = string.Empty;
-                configGuiContent.tooltip = tooltip;
-                GUI.Label(tableRowRect, configGuiContent, GUIStyle.none);
-
-                DrawConfigRow(tableRowRect, row, isHovered, isLocked);
-
-                // Draw column separators
-                if (Event.current.type == EventType.Repaint)
-                {
-                    DrawColumnSeparators(tableRowRect);
-                }
-
-                rowIndex++;
-            }
-
-            GUILayout.EndScrollView();
-        }
-
-        private void DrawColumnMenu(Rect menuRect)
-        {
-            InitializeColumnVisibility();
-
-            // Column names
-            string[] columnNames = {
-                "Name", "Thrust", "Min%", "ISP", "Mass", "Gimbal",
-                "Ignitions", "Ullage", "Press-Fed", "Rated (s)", "Tested (s)",
-                "Ign No Data", "Ign Max Data", "Burn No Data", "Burn Max Data",
-                "Tech", "Cost", "Actions"
-            };
-
-            float yPos = menuRect.y + 10;
-            float leftX = menuRect.x + 10;
-            float rightX = menuRect.x + menuRect.width / 2 + 5;
-
-            // Use cached menu styles
-            GUIStyle headerStyle = EngineConfigStyles.MenuHeader;
-            GUIStyle labelStyle = EngineConfigStyles.MenuLabel;
-
-            // Title
-            GUI.Label(new Rect(leftX, yPos, menuRect.width - 20, 20), "Column Visibility", headerStyle);
-            yPos += 25;
-
-            // Separator
-            if (Event.current.type == EventType.Repaint)
-            {
-                Texture2D separatorTex = Styles.CreateColorPixel(new Color(0.3f, 0.3f, 0.3f, 0.5f));
-                GUI.DrawTexture(new Rect(leftX, yPos, menuRect.width - 20, 1), separatorTex);
-            }
-            yPos += 10;
-
-            // Headers for Full and Compact
-            GUI.Label(new Rect(leftX + 100, yPos, 60, 20), "Full", headerStyle);
-            GUI.Label(new Rect(leftX + 170, yPos, 60, 20), "Compact", headerStyle);
-            yPos += 25;
-
-            // Scrollable area for columns
-            Rect scrollRect = new Rect(leftX, yPos, menuRect.width - 20, menuRect.height - 80);
-            Rect viewRect = new Rect(0, 0, scrollRect.width - 20, columnNames.Length * 25);
-
-            GUI.BeginGroup(scrollRect);
-            float itemY = 0;
-
-            for (int i = 0; i < columnNames.Length; i++)
-            {
-                // Column name
-                GUI.Label(new Rect(5, itemY, 90, 20), columnNames[i], labelStyle);
-
-                // Full view checkbox
-                bool newFullVisible = GUI.Toggle(new Rect(105, itemY, 20, 20), columnsVisibleFull[i], "");
-                if (newFullVisible != columnsVisibleFull[i])
-                {
-                    columnsVisibleFull[i] = newFullVisible;
-                }
-
-                // Compact view checkbox
-                bool newCompactVisible = GUI.Toggle(new Rect(175, itemY, 20, 20), columnsVisibleCompact[i], "");
-                if (newCompactVisible != columnsVisibleCompact[i])
-                {
-                    columnsVisibleCompact[i] = newCompactVisible;
-                }
-
-                itemY += 25;
-            }
-
-            GUI.EndGroup();
-
-            // Close button
-            if (GUI.Button(new Rect(menuRect.x + menuRect.width - 60, menuRect.y + menuRect.height - 30, 50, 20), "Close"))
-            {
-                showColumnMenu = false;
-            }
-        }
-
-        private void DrawHeaderRow(Rect headerRect)
-        {
-            float currentX = headerRect.x;
-            if (IsColumnVisible(0)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[0], headerRect.height),
-                    "Name", "Configuration name");
-                currentX += ConfigColumnWidths[0];
-            }
-            if (IsColumnVisible(1)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[1], headerRect.height),
-                    Localizer.GetStringByTag("#RF_EngineRF_Thrust"), "Rated thrust");
-                currentX += ConfigColumnWidths[1];
-            }
-            if (IsColumnVisible(2)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[2], headerRect.height),
-                    "Min%", "Minimum throttle");
-                currentX += ConfigColumnWidths[2];
-            }
-            if (IsColumnVisible(3)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[3], headerRect.height),
-                    Localizer.GetStringByTag("#RF_Engine_Isp"), "Sea level and vacuum Isp");
-                currentX += ConfigColumnWidths[3];
-            }
-            if (IsColumnVisible(4)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[4], headerRect.height),
-                    Localizer.GetStringByTag("#RF_Engine_Enginemass"), "Engine mass");
-                currentX += ConfigColumnWidths[4];
-            }
-            if (IsColumnVisible(5)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[5], headerRect.height),
-                    Localizer.GetStringByTag("#RF_Engine_TLTInfo_Gimbal"), "Gimbal range");
-                currentX += ConfigColumnWidths[5];
-            }
-            if (IsColumnVisible(6)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[6], headerRect.height),
-                    Localizer.GetStringByTag("#RF_EngineRF_Ignitions"), "Ignitions");
-                currentX += ConfigColumnWidths[6];
-            }
-            if (IsColumnVisible(7)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[7], headerRect.height),
-                    Localizer.GetStringByTag("#RF_Engine_ullage"), "Ullage requirement");
-                currentX += ConfigColumnWidths[7];
-            }
-            if (IsColumnVisible(8)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[8], headerRect.height),
-                    Localizer.GetStringByTag("#RF_Engine_pressureFed"), "Pressure-fed");
-                currentX += ConfigColumnWidths[8];
-            }
-            if (IsColumnVisible(9)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[9], headerRect.height),
-                    "Rated (s)", "Rated burn time");
-                currentX += ConfigColumnWidths[9];
-            }
-            if (IsColumnVisible(10)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[10], headerRect.height),
-                    "Tested (s)", "Tested burn time (real-world test duration)");
-                currentX += ConfigColumnWidths[10];
-            }
-            if (IsColumnVisible(11)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[11], headerRect.height),
-                    "Ign No Data", "Ignition reliability at 0 data");
-                currentX += ConfigColumnWidths[11];
-            }
-            if (IsColumnVisible(12)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[12], headerRect.height),
-                    "Ign Max Data", "Ignition reliability at max data");
-                currentX += ConfigColumnWidths[12];
-            }
-            if (IsColumnVisible(13)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[13], headerRect.height),
-                    "Burn No Data", "Cycle reliability at 0 data");
-                currentX += ConfigColumnWidths[13];
-            }
-            if (IsColumnVisible(14)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[14], headerRect.height),
-                    "Burn Max Data", "Cycle reliability at max data");
-                currentX += ConfigColumnWidths[14];
-            }
-            if (IsColumnVisible(15)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[15], headerRect.height),
-                    Localizer.GetStringByTag("#RF_Engine_Requires"), "Required technology");
-                currentX += ConfigColumnWidths[15];
-            }
-            if (IsColumnVisible(16)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[16], headerRect.height),
-                    "Extra Cost", "Extra cost for this config");
-                currentX += ConfigColumnWidths[16];
-            }
-            if (IsColumnVisible(17)) {
-                DrawHeaderCell(new Rect(currentX, headerRect.y, ConfigColumnWidths[17], headerRect.height),
-                    "", "Switch and purchase actions"); // No label, just tooltip
-            }
-        }
-
-        private void DrawColumnSeparators(Rect rowRect)
-        {
-            float currentX = rowRect.x;
-            for (int i = 0; i < ConfigColumnWidths.Length - 1; i++)
-            {
-                if (IsColumnVisible(i))
-                {
-                    currentX += ConfigColumnWidths[i];
-                    Rect separatorRect = new Rect(currentX, rowRect.y, 1, rowRect.height);
-                    GUI.DrawTexture(separatorRect, Textures.ColumnSeparator);
-                }
-            }
-        }
-
-        private void DrawHeaderCell(Rect rect, string text, string tooltip)
-        {
-            bool hover = rect.Contains(Event.current.mousePosition);
-            GUIStyle headerStyle = hover ? EngineConfigStyles.HeaderCellHover : EngineConfigStyles.HeaderCell;
-
-            if (configGuiContent == null)
-                configGuiContent = new GUIContent();
-            configGuiContent.text = text;
-            configGuiContent.tooltip = tooltip;
-            Matrix4x4 matrixBackup = GUI.matrix;
-            // Start text at horizontal center of column
-            float offsetX = rect.width / 2f;
-            Vector2 pivot = new Vector2(rect.x + offsetX, rect.y + rect.height + 4f);
-            GUIUtility.RotateAroundPivot(-45f, pivot);
-            GUI.Label(new Rect(rect.x + offsetX, rect.y + rect.height - 22f, 140f, 24f), configGuiContent, headerStyle);
-            GUI.matrix = matrixBackup;
-        }
-
-        private void DrawConfigRow(Rect rowRect, ConfigRowDefinition row, bool isHovered, bool isLocked)
-        {
-            // Use cached styles instead of creating new ones every frame
-            GUIStyle primaryStyle;
-            if (isLocked)
-                primaryStyle = EngineConfigStyles.RowPrimaryLocked;
-            else if (isHovered)
-                primaryStyle = EngineConfigStyles.RowPrimaryHover;
-            else
-                primaryStyle = EngineConfigStyles.RowPrimary;
-
-            GUIStyle secondaryStyle = EngineConfigStyles.RowSecondary;
-
-            float currentX = rowRect.x;
-            string nameText = row.DisplayName;
-            if (row.Indent) nameText = "    ↳ " + nameText;
-
-            if (IsColumnVisible(0)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[0], rowRect.height), nameText, primaryStyle);
-                currentX += ConfigColumnWidths[0];
-            }
-
-            if (IsColumnVisible(1)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[1], rowRect.height), GetThrustString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[1];
-            }
-
-            if (IsColumnVisible(2)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[2], rowRect.height), GetMinThrottleString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[2];
-            }
-
-            if (IsColumnVisible(3)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[3], rowRect.height), GetIspString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[3];
-            }
-
-            if (IsColumnVisible(4)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[4], rowRect.height), GetMassString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[4];
-            }
-
-            if (IsColumnVisible(5)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[5], rowRect.height), GetGimbalString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[5];
-            }
-
-            if (IsColumnVisible(6)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[6], rowRect.height), GetIgnitionsString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[6];
-            }
-
-            if (IsColumnVisible(7)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[7], rowRect.height), GetBoolSymbol(row.Node, "ullage"), secondaryStyle);
-                currentX += ConfigColumnWidths[7];
-            }
-
-            if (IsColumnVisible(8)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[8], rowRect.height), GetBoolSymbol(row.Node, "pressureFed"), secondaryStyle);
-                currentX += ConfigColumnWidths[8];
-            }
-
-            if (IsColumnVisible(9)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[9], rowRect.height), GetRatedBurnTimeString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[9];
-            }
-
-            if (IsColumnVisible(10)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[10], rowRect.height), GetTestedBurnTimeString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[10];
-            }
-
-            if (IsColumnVisible(11)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[11], rowRect.height), GetIgnitionReliabilityStartString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[11];
-            }
-
-            if (IsColumnVisible(12)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[12], rowRect.height), GetIgnitionReliabilityEndString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[12];
-            }
-
-            if (IsColumnVisible(13)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[13], rowRect.height), GetCycleReliabilityStartString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[13];
-            }
-
-            if (IsColumnVisible(14)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[14], rowRect.height), GetCycleReliabilityEndString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[14];
-            }
-
-            if (IsColumnVisible(15)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[15], rowRect.height), GetTechString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[15];
-            }
-
-            if (IsColumnVisible(16)) {
-                GUI.Label(new Rect(currentX, rowRect.y, ConfigColumnWidths[16], rowRect.height), GetCostDeltaString(row.Node), secondaryStyle);
-                currentX += ConfigColumnWidths[16];
-            }
-
-            if (IsColumnVisible(17)) {
-                DrawActionCell(new Rect(currentX, rowRect.y + 1, ConfigColumnWidths[17], rowRect.height - 2), row.Node, row.IsSelected, row.Apply);
-            }
-        }
-
-        private void DrawActionCell(Rect rect, ConfigNode node, bool isSelected, Action apply)
-        {
-            GUIStyle smallButtonStyle = EngineConfigStyles.SmallButton;
-
-            string configName = node.GetValue("name");
-            bool canUse = EngineConfigTechLevels.CanConfig(node);
-            bool unlocked = EngineConfigTechLevels.UnlockedConfig(node, part);
-            double cost = EntryCostManager.Instance.ConfigEntryCost(configName);
-
-            // Auto-purchase free configs
-            if (cost <= 0 && !unlocked && canUse)
-                EntryCostManager.Instance.PurchaseConfig(configName, node.GetValue("techRequired"));
-
-            // Split the rect into two buttons side by side
-            float buttonWidth = rect.width / 2f - 2f;
-            Rect switchRect = new Rect(rect.x, rect.y, buttonWidth, rect.height);
-            Rect purchaseRect = new Rect(rect.x + buttonWidth + 4f, rect.y, buttonWidth, rect.height);
-
-            // Switch button - always enabled except when already selected
-            GUI.enabled = !isSelected;
-            string switchLabel = isSelected ? "Active" : "Switch";
-            if (GUI.Button(switchRect, switchLabel, smallButtonStyle))
-            {
-                if (!unlocked && cost <= 0)
-                    EntryCostManager.Instance.PurchaseConfig(configName, node.GetValue("techRequired"));
-                apply?.Invoke();
-            }
-
-            // Purchase button (shows cost)
-            GUI.enabled = canUse && !unlocked && cost > 0;
-            string purchaseLabel;
-            if (cost > 0)
-                purchaseLabel = unlocked ? "Owned" : $"Buy ({cost:N0}√)";
-            else
-                purchaseLabel = "Free";
-
-            if (GUI.Button(purchaseRect, purchaseLabel, smallButtonStyle))
-            {
-                if (EntryCostManager.Instance.PurchaseConfig(configName, node.GetValue("techRequired")))
-                    apply?.Invoke();
-            }
-
-            GUI.enabled = true;
-        }
-
-        internal string GetThrustString(ConfigNode node)
-        {
-            if (!node.HasValue(thrustRating))
-                return "-";
-
-            float thrust = scale * TechLevels.ThrustTL(node.GetValue(thrustRating), node);
-            // Remove decimals for large thrust values
-            if (thrust >= 100f)
-                return $"{thrust:N0} kN";
-            return $"{thrust:N2} kN";
-        }
-
-        internal string GetMinThrottleString(ConfigNode node)
-        {
-            float value = -1f;
-            if (node.HasValue("minThrust") && node.HasValue(thrustRating))
-            {
-                float.TryParse(node.GetValue("minThrust"), out float minT);
-                float.TryParse(node.GetValue(thrustRating), out float maxT);
-                if (maxT > 0)
-                    value = minT / maxT;
-            }
-            else if (node.HasValue("throttle"))
-            {
-                float.TryParse(node.GetValue("throttle"), out value);
-            }
-
-            if (value < 0f)
-                return "-";
-            return value.ToString("P0");
-        }
-
-        internal string GetIspString(ConfigNode node)
-        {
-            if (node.HasNode("atmosphereCurve"))
-            {
-                FloatCurve isp = new FloatCurve();
-                isp.Load(node.GetNode("atmosphereCurve"));
-                float ispVac = isp.Evaluate(isp.maxTime);
-                float ispSL = isp.Evaluate(isp.minTime);
-                return $"{ispVac:N0}-{ispSL:N0}";
-            }
-
-            if (node.HasValue("IspSL") && node.HasValue("IspV"))
-            {
-                float.TryParse(node.GetValue("IspSL"), out float ispSL);
-                float.TryParse(node.GetValue("IspV"), out float ispV);
-                if (techLevel != -1)
-                {
-                    TechLevel cTL = new TechLevel();
-                    if (cTL.Load(node, techNodes, engineType, techLevel))
-                    {
-                        ispSL *= ispSLMult * cTL.AtmosphereCurve.Evaluate(1);
-                        ispV *= ispVMult * cTL.AtmosphereCurve.Evaluate(0);
-                    }
-                }
-                return $"{ispV:N0}-{ispSL:N0}";
-            }
-
-            return "-";
-        }
-
-        internal string GetMassString(ConfigNode node)
-        {
-            if (origMass <= 0f)
-                return "-";
-
-            float cMass = scale * origMass * RFSettings.Instance.EngineMassMultiplier;
-            if (node.HasValue("massMult") && float.TryParse(node.GetValue("massMult"), out float ftmp))
-                cMass *= ftmp;
-
-            return $"{cMass:N3}t";
-        }
-
-        internal string GetGimbalString(ConfigNode node)
-        {
-            if (!part.HasModuleImplementing<ModuleGimbal>())
-                return "<color=#9E9E9E>✗</color>";
-
-            var gimbals = ExtractGimbals(node);
-
-            // If no explicit gimbal in config, check if we should use tech level gimbal
-            if (gimbals.Count == 0 && techLevel != -1 && (!gimbalTransform.Equals(string.Empty) || useGimbalAnyway))
-            {
-                TechLevel cTL = new TechLevel();
-                if (cTL.Load(node, techNodes, engineType, techLevel))
-                {
-                    float gimbalRange = cTL.GimbalRange;
-                    if (node.HasValue("gimbalMult"))
-                        gimbalRange *= float.Parse(node.GetValue("gimbalMult"), CultureInfo.InvariantCulture);
-
-                    if (gimbalRange >= 0)
-                        return $"{gimbalRange * gimbalMult:0.#}°";
-                }
-            }
-
-            // Fallback: if config has no gimbal data, use the part's ModuleGimbal
-            if (gimbals.Count == 0)
-            {
-                foreach (var gimbalMod in part.Modules.OfType<ModuleGimbal>())
-                {
-                    if (gimbalMod != null)
-                    {
-                        var gimbal = new Gimbal(gimbalMod.gimbalRange, gimbalMod.gimbalRange, gimbalMod.gimbalRange, gimbalMod.gimbalRange, gimbalMod.gimbalRange);
-                        gimbals[gimbalMod.gimbalTransformName] = gimbal;
-                    }
-                }
-            }
-
-            if (gimbals.Count == 0)
-                return "<color=#9E9E9E>✗</color>";
-
-            var first = gimbals.Values.First();
-            bool allSame = gimbals.Values.All(g => g.gimbalRange == first.gimbalRange
-                && g.gimbalRangeXP == first.gimbalRangeXP
-                && g.gimbalRangeXN == first.gimbalRangeXN
-                && g.gimbalRangeYP == first.gimbalRangeYP
-                && g.gimbalRangeYN == first.gimbalRangeYN);
-
-            if (allSame)
-                return first.Info();
-
-            // Multiple different gimbal ranges - list them all
-            var uniqueInfos = gimbals.Values.Select(g => g.Info()).Distinct().OrderBy(s => s);
-            return string.Join(", ", uniqueInfos);
-        }
-
-        internal string GetIgnitionsString(ConfigNode node)
-        {
-            if (!node.HasValue("ignitions"))
-                return "-";
-
-            if (!int.TryParse(node.GetValue("ignitions"), out int ignitions))
-                return "∞";
-
-            int resolved = TechLevels.ConfigIgnitions(ignitions);
-            if (resolved == -1)
-                return "∞";
-            if (resolved == 0 && literalZeroIgnitions)
-                return "<color=#FFEB3B>Gnd</color>"; // Yellow G for ground-only ignitions
-            return resolved.ToString();
-        }
-
-        internal string GetBoolSymbol(ConfigNode node, string key)
-        {
-            if (!node.HasValue(key))
-                return "<color=#9E9E9E>✗</color>"; // Treat missing as false - gray (no restriction)
-            bool isTrue = node.GetValue(key).ToLower() == "true";
-            return isTrue ? "<color=#FFA726>✓</color>" : "<color=#9E9E9E>✗</color>"; // Orange for restriction, gray for no restriction
-        }
-
-        private void InitializeColumnVisibility()
-        {
-            if (columnVisibilityInitialized)
-                return;
-
-            // Initialize full view: all columns visible by default
-            for (int i = 0; i < 18; i++)
-                columnsVisibleFull[i] = true;
-
-            // Initialize compact view: only essential columns
-            for (int i = 0; i < 18; i++)
-                columnsVisibleCompact[i] = false;
-
-            // Essential columns for compact view
-            int[] compactColumns = { 0, 1, 3, 4, 6, 9, 10, 15, 16, 17 }; // Tech, Cost, Actions
-            foreach (int col in compactColumns)
-                columnsVisibleCompact[col] = true;
-
-            columnVisibilityInitialized = true;
-        }
-
-        private bool IsColumnVisible(int columnIndex)
-        {
-            InitializeColumnVisibility();
-
-            if (columnIndex < 0 || columnIndex >= 18)
-                return false;
-
-            return compactView ? columnsVisibleCompact[columnIndex] : columnsVisibleFull[columnIndex];
-        }
-
-        internal string GetRatedBurnTimeString(ConfigNode node)
-        {
-            bool hasRatedBurnTime = node.HasValue("ratedBurnTime");
-            bool hasRatedContinuousBurnTime = node.HasValue("ratedContinuousBurnTime");
-
-            if (!hasRatedBurnTime && !hasRatedContinuousBurnTime)
-                return "∞";
-
-            // If both values exist, show as "continuous/cumulative"
-            if (hasRatedBurnTime && hasRatedContinuousBurnTime)
-            {
-                string continuous = node.GetValue("ratedContinuousBurnTime");
-                string cumulative = node.GetValue("ratedBurnTime");
-                return $"{continuous}/{cumulative}";
-            }
-
-            // Otherwise show whichever one exists
-            return hasRatedBurnTime ? node.GetValue("ratedBurnTime") : node.GetValue("ratedContinuousBurnTime");
-        }
-
-        internal string GetTestedBurnTimeString(ConfigNode node)
-        {
-            // Values are copied to CONFIG level by ModuleManager patch
-            if (!node.HasValue("testedBurnTime"))
-                return "-";
-
-            float testedBurnTime = 0f;
-            if (node.TryGetValue("testedBurnTime", ref testedBurnTime))
-                return testedBurnTime.ToString("F0");
-
-            return "-";
-        }
-
-        internal string GetIgnitionReliabilityStartString(ConfigNode node)
-        {
-            // Values are copied to CONFIG level by ModuleManager patch
-            if (!node.HasValue("ignitionReliabilityStart"))
-                return "-";
-            if (float.TryParse(node.GetValue("ignitionReliabilityStart"), out float val))
-                return $"{val:P1}";
-            return "-";
-        }
-
-        internal string GetIgnitionReliabilityEndString(ConfigNode node)
-        {
-            // Values are copied to CONFIG level by ModuleManager patch
-            if (!node.HasValue("ignitionReliabilityEnd"))
-                return "-";
-            if (float.TryParse(node.GetValue("ignitionReliabilityEnd"), out float val))
-                return $"{val:P1}";
-            return "-";
-        }
-
-        internal string GetCycleReliabilityStartString(ConfigNode node)
-        {
-            // Values are copied to CONFIG level by ModuleManager patch
-            if (!node.HasValue("cycleReliabilityStart"))
-                return "-";
-            if (float.TryParse(node.GetValue("cycleReliabilityStart"), out float val))
-                return $"{val:P1}";
-            return "-";
-        }
-
-        internal string GetCycleReliabilityEndString(ConfigNode node)
-        {
-            // Values are copied to CONFIG level by ModuleManager patch
-            if (!node.HasValue("cycleReliabilityEnd"))
-                return "-";
-            if (float.TryParse(node.GetValue("cycleReliabilityEnd"), out float val))
-                return $"{val:P1}";
-            return "-";
-        }
-
-        private string GetFlightDataString()
-        {
-            // Get current flight data from TestFlight
-            float currentData = TestFlightWrapper.GetCurrentFlightData(part);
-            float maxData = TestFlightWrapper.GetMaximumData(part);
-
-            if (currentData < 0f || maxData <= 0f)
-                return "-";
-
-            return $"{currentData:F0} / {maxData:F0}";
-        }
-
-        internal string GetTechString(ConfigNode node)
-        {
-            if (!node.HasValue("techRequired"))
-                return "-";
-
-            string tech = node.GetValue("techRequired");
-            if (techNameToTitle.TryGetValue(tech, out string title))
-                tech = title;
-
-            // Abbreviate: keep first word, then first 4 letters of other words with "-"
-            var words = tech.Split(' ');
-            if (words.Length <= 1)
-                return tech;
-
-            var abbreviated = words[0];
-            for (int i = 1; i < words.Length; i++)
-            {
-                if (words[i].Length > 4)
-                    abbreviated += "-" + words[i].Substring(0, 4);
-                else
-                    abbreviated += "-" + words[i];
-            }
-            return abbreviated;
-        }
-
-        internal string GetCostDeltaString(ConfigNode node)
-        {
-            if (!node.HasValue("cost"))
-                return "-";
-
-            float curCost = scale * float.Parse(node.GetValue("cost"), CultureInfo.InvariantCulture);
-            if (techLevel != -1)
-                curCost = TechLevels.CostTL(curCost, node) - TechLevels.CostTL(0f, node);
-
-            if (Mathf.Approximately(curCost, 0f))
-                return "-";
-
-            string sign = curCost < 0 ? string.Empty : "+";
-            return $"{sign}{curCost:N0}√";
-        }
-
-        #region Removed TestFlight UI Integration
-        // All TestFlight column display code removed due to:
-        // 1. Data spread across multiple modules (TestFlightCore, TestFlightFailure_IgnitionFail)
-        // 2. Reliability/MTBF values require complex calculations, not simple field access
-        // 3. Reflection-based approach was error-prone and caused GUI crashes
-        //
-        // TestFlight integration still works via UpdateTFInterops() to notify TestFlight
-        // of active configuration changes. TestFlight UI displays its own data.
-        //
-        // Removed methods:
-        // - GetFlightDataString, GetIgnitionChanceString, GetIgnitionChanceAtMaxDataString
-        // - GetReliabilityString, GetReliabilityAtMaxDataString
-        // - TryGetTestFlightStats, GetAllTestFlightDataSources, GetTestFlightDataSource
-        // - TryGetConfigDataSource, TryGetNumber, TryGetMemberValue, TryGetStringMember
-        // - TryConvertToDouble, FormatPercent
-        // - TestFlightStats struct
-        #endregion
-
-        private string GetRowTooltip(ConfigNode node)
-        {
-            List<string> tooltipParts = new List<string>();
-
-            // Add description if present
-            if (node.HasValue("description"))
-                tooltipParts.Add(node.GetValue("description"));
-
-            // Add propellants with flow rates if present
-            if (node.HasNode("PROPELLANT"))
-            {
-                // Get thrust and ISP for flow calculations
-                float thrust = 0f;
-                float isp = 0f;
-
-                if (node.HasValue(thrustRating) && float.TryParse(node.GetValue(thrustRating), out float maxThrust))
-                    thrust = TechLevels.ThrustTL(node.GetValue(thrustRating), node) * scale;
-
-                if (node.HasNode("atmosphereCurve"))
-                {
-                    var atmCurve = new FloatCurve();
-                    atmCurve.Load(node.GetNode("atmosphereCurve"));
-                    isp = atmCurve.Evaluate(0f); // Vacuum ISP
-                }
-
-                // Calculate total mass flow: F = mdot * Isp * g0
-                // Thrust is in kN (kilonewtons), convert to N (newtons) for the equation
-                const float g0 = 9.80665f;
-                float thrustN = thrust * 1000f;
-                float totalMassFlow = (thrustN > 0f && isp > 0f) ? thrustN / (isp * g0) : 0f;
-
-                // Get propellant ratios
-                var propNodes = node.GetNodes("PROPELLANT");
-                float totalRatio = 0f;
-                foreach (var propNode in propNodes)
-                {
-                    string ratioStr = null;
-                    if (propNode.TryGetValue("ratio", ref ratioStr) && float.TryParse(ratioStr, out float ratio))
-                        totalRatio += ratio;
-                }
-
-                var propellantLines = new List<string>();
-                foreach (var propNode in propNodes)
-                {
-                    string name = propNode.GetValue("name");
-                    if (string.IsNullOrWhiteSpace(name)) continue;
-
-                    string line = $"  • {name}";
-
-                    string ratioStr2 = null;
-                    if (propNode.TryGetValue("ratio", ref ratioStr2) && float.TryParse(ratioStr2, out float ratio) && totalMassFlow > 0f && totalRatio > 0f)
-                    {
-                        float propMassFlow = totalMassFlow * (ratio / totalRatio);
-
-                        // Get density from resource library
-                        var resource = PartResourceLibrary.Instance?.GetDefinition(name);
-
-                        // Format mass flow: use grams if < 1 kg/s for better precision
-                        string massFlowStr = propMassFlow >= 1f
-                            ? $"{propMassFlow:F2} kg/s"
-                            : $"{propMassFlow * 1000f:F1} g/s";
-
-                        if (resource != null)
-                        {
-                            float volumeFlow = propMassFlow / (float)resource.density;
-                            line += $": {volumeFlow:F2} L/s ({massFlowStr})";
-                        }
-                        else
-                        {
-                            line += $": {massFlowStr}";
-                        }
-                    }
-
-                    propellantLines.Add(line);
-                }
-
-                if (propellantLines.Count > 0)
-                    tooltipParts.Add($"<b>Propellant Consumption:</b>\n{string.Join("\n", propellantLines)}");
-            }
-
-            return tooltipParts.Count > 0 ? string.Join("\n\n", tooltipParts) : string.Empty;
+            // Default implementation - just draw the table
+            // Derived classes can override to add custom UI
         }
 
         /// <summary>
-        /// Ensures textures are initialized. Handles Unity texture destruction on scene changes.
+        /// Internal callback for GUI to apply a selected configuration.
         /// </summary>
-        private void EnsureTexturesAndStyles()
+        internal void GUIApplyConfig(string configName)
         {
-            Textures.EnsureInitialized();
-            EngineConfigStyles.Initialize();
+            SetConfiguration(configName);
+            UpdateSymmetryCounterparts();
         }
 
-        virtual protected void DrawConfigSelectors(IEnumerable<ConfigNode> availableConfigNodes)
+        /// <summary>
+        /// Hook point for external mod compatibility (e.g., RP-1 Harmony patches).
+        /// Called by the GUI before rendering each config row to allow external mods
+        /// to track context via Harmony prefix/postfix patches.
+        /// Does not render anything - rendering is handled by EngineConfigGUI.
+        /// </summary>
+        internal void DrawSelectButton(ConfigNode node, bool isSelected, Action<string> applyCallback)
         {
-            DrawConfigTable(BuildConfigRows());
+            // Hook point only - external mods (RP-1) patch this to track tech node context
+            // Actual rendering is handled by EngineConfigGUI table system
         }
 
-        private void EngineManagerGUI(int WindowID)
+        public void OnGUI()
         {
-            // Use BeginVertical with GUILayout.ExpandHeight(false) to prevent extra vertical space
-            GUILayout.BeginVertical(GUILayout.ExpandHeight(false));
-
-            GUILayout.Space(4); // Minimal top padding
-
-            GUILayout.BeginHorizontal();
-            GUILayout.Label(EditorDescription);
-            GUILayout.FlexibleSpace();
-            if (GUILayout.Button(compactView ? "Full View" : "Compact View", GUILayout.Width(100)))
-            {
-                compactView = !compactView;
-            }
-            if (GUILayout.Button("Settings", GUILayout.Width(70)))
-            {
-                showColumnMenu = !showColumnMenu;
-            }
-            GUILayout.EndHorizontal();
-
-            GUILayout.Space(4); // Minimal space before table
-            DrawConfigSelectors(FilteredDisplayConfigs(false));
-
-            // Draw failure probability chart for current config
-            if (config != null && config.HasValue("cycleReliabilityStart"))
-            {
-                GUILayout.Space(6);
-
-                // Update chart settings from instance fields
-                Chart.UseLogScaleX = useLogScaleX;
-                Chart.UseLogScaleY = useLogScaleY;
-                Chart.UseSimulatedData = useSimulatedData;
-                Chart.SimulatedDataValue = simulatedDataValue;
-                Chart.ClusterSize = clusterSize;
-                Chart.ClusterSizeInput = clusterSizeInput;
-                Chart.DataValueInput = dataValueInput;
-
-                // Draw the chart
-                Chart.Draw(config, guiWindowRect.width - 10, 360);
-
-                // Update instance fields from chart (for UI controls)
-                useLogScaleX = Chart.UseLogScaleX;
-                useLogScaleY = Chart.UseLogScaleY;
-                useSimulatedData = Chart.UseSimulatedData;
-                simulatedDataValue = Chart.SimulatedDataValue;
-                clusterSize = Chart.ClusterSize;
-                clusterSizeInput = Chart.ClusterSizeInput;
-                dataValueInput = Chart.DataValueInput;
-
-                GUILayout.Space(6); // Consistent small space after chart
-            }
-
-            TechLevels.DrawTechLevelSelector();
-
-            GUILayout.Space(4); // Minimal bottom padding
-
-            GUILayout.EndVertical();
-
-            if (!myToolTip.Equals(string.Empty) && GUI.tooltip.Equals(string.Empty))
-            {
-                if (counterTT > 4)
-                {
-                    myToolTip = GUI.tooltip;
-                    counterTT = 0;
-                }
-                else
-                {
-                    counterTT++;
-                }
-            }
-            else
-            {
-                myToolTip = GUI.tooltip;
-                counterTT = 0;
-            }
-
-            GUI.DragWindow();
+            if (isMaster && HighLogic.LoadedSceneIsEditor)
+                GUI.OnGUI();
         }
+
 
         #endregion
