+++ Source/Engines/UI/ChartMath.cs
